feat: build components with tailwind

+++ packages/ui/utils/build/rollup.config.js
@@ -1,12 +1,14 @@
 import resolve from "@rollup/plugin-node-resolve"; // 告诉 Rollup 如何查找外部模块
 import commonjs from "@rollup/plugin-commonjs"; // 将CommonJS模块转换为 ES2015 供 Rollup 处理
-import esbuild from 'rollup-plugin-esbuild'
+import esbuild from "rollup-plugin-esbuild";
 import vue from "rollup-plugin-vue"; // 处理vue文件
 import css from "rollup-plugin-css-only"; // 提取css，压缩能力不行
+import postcss from "rollup-plugin-postcss";
 import CleanCSS from "clean-css"; // 压缩css
 import fs from "fs-extra"; // 写文件
 
 export default fs.readdirSync("packages").map((name) => {
+  if (name === "theme") return null;
   if (fs.existsSync(`./packages/${name}/lib`)) {
     fs.emptyDirSync(`./packages/${name}/lib`);
   } else {
@@ -18,6 +20,15 @@ export default fs.readdirSync("packages").map((name) => {
     plugins: [
       resolve({ extensions: [".vue"] }),
       commonjs(),
+      postcss({
+        config: {
+          path: "./postcss.config.js",
+        },
+        extensions: [".css"],
+        extract: true,
+        // minimize: process.env.NODE_ENV === "production",
+        // modules: true,
+      }),
       css({
         output(style) {
           // 压缩 css 写入 dist/vue-rollup-component-template.css
@@ -29,7 +40,7 @@ export default fs.readdirSync("packages").map((name) => {
       }),
       // css: false 将<style>块转换为导入语句，rollup-plugin-css-only可以提取.vue文件中的样式
       vue(),
-      esbuild()
+      esbuild(),
     ],
     output: {
       name: "index",
